+++ package.json
@@ -39,7 +39,6 @@
     "gatsby-plugin-remove-serviceworker": "^1.0.0",
     "gatsby-plugin-sharp": "^2.2.28",
     "gatsby-plugin-sitemap": "^2.2.16",
-    "gatsby-remark-autolink-headers": "^2.1.13",
     "gatsby-remark-copy-linked-files": "^2.1.25",
     "gatsby-remark-images": "^3.1.25",
     "gatsby-remark-prismjs": "^3.3.17",
@@ -48,6 +47,7 @@
     "gatsby-source-filesystem": "^2.1.30",
     "gatsby-transformer-json": "^2.2.12",
     "gatsby-transformer-sharp": "^2.2.20",
+    "github-slugger": "^1.2.1",
     "lowmess-prism": "^0.2.3",
     "prismjs": "^1.17.1",
     "prop-types": "^15.7.2",
